feat(ws-14): consolidate error engine to error/engine (COMPLETE)
- Created error/engine/ directory structure
- Moved 8 error modules from 2 locations to error/engine/:
  From src/pipeline/: error_context.py, error_engine.py, error_pipeline_cli.py,
                      error_pipeline_service.py, error_state_machine.py
  From error/: plugin_manager.py, pipeline_engine.py, file_hash_cache.py
- Created shims in src/pipeline/ for backward compatibility
- Created shims in error/ for error engine files
- Fixed imports (removed stub AgentCoordinator dependency)

Directory consolidation complete:
✓ error/shared/utils/ (6 utility modules)
✓ error/plugins/ (21 plugins)
✓ error/engine/ (8 engine modules)

Acceptance tests:
✓ from error.engine.error_engine import run_error_pipeline
✓ from error.engine.plugin_manager import PluginManager
✓ python scripts/run_error_engine.py --help

Risk: HIGH
Status: COMPLETE - Phase C (Error Subsystem) 100% COMPLETE!
