+++ error/engine/__init__.py
@@ -1,0 +1,15 @@
+"""Error Engine
+
+Error pipeline orchestration, plugin management, and state machine.
+"""
+
+__all__ = [
+    "plugin_manager",
+    "pipeline_engine", 
+    "file_hash_cache",
+    "error_context",
+    "error_engine",
+    "error_state_machine",
+    "error_pipeline_cli",
+    "error_pipeline_service"
+]
